refactor(cli): move wrapText to lib/utils
